adjust resource group api
